old/backend/app/uvaPR_Microservices/uvaPR_Microservices/requirements.txt
Django==1.10
djangorestframework==3.7.7
elasticsearch==2.4.1
kafka-python==1.0.0
mod-wsgi==4.5.7
mysqlclient==1.3.8
Pillow==5.0.0
six==1.10.0
urllib3==1.20
django-cors-headers==2.1.0